Term life insurance data
We looked at how different things like age, income, and gender might affect the amount a company pays in case of someone's passing (FACE). Through graphs and calculations, we tried to see if these factors have a connection with each other or with the FACE amount. The results showed that gender (being male or female) seemed to be more related to the predicted FACE amount compared to age and income. However, age and income didn't appear to strongly impact the predicted FACE amount in this analysis. we found that being male or female might be somewhat linked to the amount paid, but age and income may not play as big of a role in determining that amount in this particular scenario.